Upgrade to GHC 8.8.3 and remove some unused deps/features
