+++ env.yaml.example
@@ -3,7 +3,6 @@ web:
   port: 1337
   tls: false
   pretty: true
-  monitor: false
 
 ipfs:
   url: http://localhost:5001
