--- library/Fission/Internal/Orphanage/MultipartForm.hs
@@ -1,35 +1,0 @@
-{-# OPTIONS_GHC -fno-warn-orphans #-}
-
-module Fission.Internal.Orphanage.MultipartForm () where
-
-import Data.Swagger
-import Network.HTTP.Media.MediaType
-import Servant
-import Servant.Multipart
-import Servant.Swagger
-import Servant.Swagger.Internal
-
-import Fission.Prelude
-
-
-instance HasSwagger api => HasSwagger (MultipartForm Mem (MultipartData Mem) :> api) where
-  toSwagger _ =
-    Proxy @api
-      |> toSwagger
-      |> addConsumes ["multipart" // "form-data"]
-      |> addParam param
-    where
-      param =
-        mempty
-          |> name        .~ "file"
-          |> required    ?~ True
-          |> description ?~ "A file to upload (may also be multipart/form-data)"
-          |> schema      .~ ParamOther fileParam
-
-      fileParam =
-        mempty
-          |> in_         .~ ParamFormData
-          |> paramSchema .~ swaggerFileType
-
-      swaggerFileType =
-        mempty |> type_ ?~ SwaggerFile
