--- library/Fission/Monitor.hs
@@ -1,29 +1,0 @@
--- | Application monitoring
-module Fission.Monitor
-  ( Config (..)
-  , wai
-  ) where
-
-import System.Envy
-import System.Remote.Monitoring.Wai
-
-import Fission.Prelude
-import Fission.Monitor.Types
-
-wai :: (MonadIO m, MonadLogger m) => m ()
-wai = liftIO decodeEnv >>= \case
-  Left err -> do
-    logError <| displayShow err
-    Config { ekgHost, ekgPort } <- return defConfig
-    liftIO <| void <| forkServer ekgHost ekgPort
-
-  Right Config { ekgHost, ekgPort } -> do
-    logInfo <| mconcat
-      [ "EKG available at "
-      , displayBytesUtf8 ekgHost
-      , ":"
-      , display ekgPort
-      , "\n"
-      ]
-
-    liftIO <| void <| forkServer ekgHost ekgPort
