+++ library/Fission/Web/Swagger.hs
@@ -13,7 +13,6 @@ import qualified Servant.Swagger.Internal.TypeLevel.API as Servant.API
 import           Fission.Prelude
 import           Fission.Internal.Orphanage.BasicAuth     ()
 import           Fission.Internal.Orphanage.BasicAuthData ()
-import           Fission.Internal.Orphanage.MultipartForm ()
 import           Fission.Internal.Orphanage.HigherOrder   ()
 import           Fission.Internal.Orphanage.RegisterDid   ()
