+++ package.yaml
@@ -92,21 +92,18 @@ dependencies:
   - bytestring
   - data-default-class
   - data-has
-  - ekg-wai
   - envy
   - esqueleto
   - exceptions
   - file-embed
   - filepath
   - flow
-  - fsnotify
   - Glob
-  - haskeline
   - http-client
   - http-client-tls
   - http-media
   - http-types
-  - ip
+  # - ip
   - ipfs
   - lens
   - lens-aeson
@@ -132,7 +129,6 @@ dependencies:
   - servant-auth-swagger
   - servant-client
   - servant-client-core
-  - servant-multipart
   - servant-server
   - servant-swagger
   - servant-swagger-ui
