+++ stack.yaml
@@ -17,7 +17,7 @@
 #
 # resolver: ./custom-snapshot.yaml
 # resolver: https://example.com/snapshots/2018-01-01.yaml
-resolver: lts-14.25
+resolver: lts-15.4
 
 # User packages to be built.
 # Various formats can be used as shown in the example below.
@@ -39,16 +39,22 @@ packages:
 # (e.g., acme-missiles-0.3)
 extra-deps:
 - alex-3.2.5
-- amazonka-1.6.1
-- amazonka-core-1.6.1
-- amazonka-route53-1.6.1
-- ekg-wai-0.1.0.3
+# - amazonka-1.6.1
+# - amazonka-core-1.6.1
+# - amazonka-route53-1.6.1
+# - Cabal-3.0.0.0
+# - directory-1.3.6.0
 - happy-1.19.12
-- ipfs-1.0.1
+#- ipfs-1.0.2@sha2-c7fd74fb4c5486ec48f446d08743e4423159baeb 
+- github: fission-suite/ipfs-haskell
+  commit: c7fd74fb4c5486ec48f446d08743e4423159baeb 
 - raven-haskell-0.1.4.0
 - servant-quickcheck-0.0.7.4
 - tasty-rerun-1.1.17
 - text-time-0.2.0
+# - ekg-core-0.1.1.7
+# - ekg-json-0.1.0.6
+# - unix-2.7.2.2
 
 # Override default flag values for local packages and extra-deps
 # flags: {}
